+++ backend/api.py
@@ -286,16 +286,23 @@ def get_cache(natural_query: str) -> dict:
         .with_limit(1)
         .do()
     )
-    return results
+
+    if natural_query == results["data"]["Get"]["CachedResult"][0]["naturalQuery"]:
+        return results
+    else:
+        return {"data": {"Get": {"CachedResult": []}}}
 
 
-def get_cache_count() -> None:
-    """Update the global cache count
-    @returns None
+def get_cache_count() -> list:
+    """Update the global cache count and return all cached queries
+    @returns list of queries
     """
-    global cache_count
     query = client.query.get("CachedResult", ["naturalQuery"]).do()
-    cache_count = len(query["data"]["Get"]["CachedResult"])
+    cachedQueries = [
+        naturalQuery["naturalQuery"]
+        for naturalQuery in query["data"]["Get"]["CachedResult"]
+    ]
+    return cachedQueries
 
 
 def check_cache(cache_results: dict, natural_query: str, max_distance: float) -> dict:
@@ -307,6 +314,10 @@ def check_cache(cache_results: dict, natural_query: str, max_distance: float) ->
     """
     if cache_results["data"]["Get"]["CachedResult"]:
         msg.good("Cache entry exists!")
+        cache_results["data"]["Get"]["CachedResult"][0]["summary"] = (
+            "🛰️ RETRIEVED FROM CACHE: "
+            + cache_results["data"]["Get"]["CachedResult"][0]["summary"]
+        )
         return cache_results
     else:
         msg.warn("Cache entry does not exist!")
@@ -332,6 +343,10 @@ def check_cache(cache_results: dict, natural_query: str, max_distance: float) ->
         else:
             msg.good(
                 f'Retrieved similar results (distance {results["data"]["Get"]["CachedResult"][0]["_additional"]["distance"]})'
+            )
+            results["data"]["Get"]["CachedResult"][0]["summary"] = (
+                f"⭐ RETURNED SIMILAR RESULTS FROM QUERY '{natural_query}' ({round(results['data']['Get']['CachedResult'][0]['_additional']['distance'],2)}) : "
+                + results["data"]["Get"]["CachedResult"][0]["summary"]
             )
             return results
 
@@ -456,22 +471,27 @@ class NLQuery(BaseModel):
 # Define health check endpoint
 @app.get("/health")
 async def root():
+    global cache_count
+
     try:
-        client.schema.get()
-        get_cache_count()
+        cached_queries = get_cache_count()
+        cache_count = len(cached_queries)
         return JSONResponse(
             content={
                 "message": "Alive!",
                 "requests": request_count,
                 "cache_count": cache_count,
+                "cache_queries": cached_queries,
             }
         )
-    except:
+    except Exception as e:
+        msg.fail(f"Healthcheck failed with {str(e)}")
         return JSONResponse(
             content={
                 "message": "Database connection failed!",
                 "requests": request_count,
                 "cache_count": cache_count,
+                "cache_queries": [],
             },
             status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
         )
@@ -606,7 +626,7 @@ async def generate_query(payload: NLQuery):
                             ]
                         ),
                         "results": results,
-                        "generative_summary": generative_summary,
+                        "generative_summary": "✨ GENERATED: " + generative_summary,
                     }
                 )